requirements.txt
# Core
httpx>=0.27.0
pydantic>=2.0.0
pyyaml>=6.0

# Voice Output - ElevenLabs
elevenlabs>=1.0.0
pygame>=2.5.0
websockets>=12.0

# Async
anyio>=4.0.0

# Utils
python-dotenv>=1.0.0
rich>=13.0.0
